+++ package.json
@@ -4,7 +4,8 @@
   "main": "main.js",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "nodemon main.js"
   },
   "dependencies": {
     "cors": "^2.8.5",
